Changes to P-type control
Changed the control logic so that the car turns whenever it detects the
black line. This ensures that it will stay within the white, assuming
the cameras are initially pointed at only white.

Also, increased the turn degree by about 200, it looks a bit
oversensitive right now.

Also formatted the original control loop to look nicer. It is still in
the code, just commented it out while i tested the new one.