Turn Speed Restored
Progress Note: slowing turn speeds does increase odds of making the turn
after the hill, but may not be worth it, as it slows the overall run a
fair amount. Current goal is to detect the hill, not for the sake of the
hill itself, but to make potential turns that happen right after the
hill.

For the time being, we can adjust the code the match to the exact track,
as the track layout is available for the weekend before the competition.
Changing this can be delayed until spring, when we won't know the exact
layout for the tournament.